+++ tasks.py
@@ -88,6 +88,6 @@ def tree(c: Context):
 @task
 def update(c: Context):
     """Update dependencies"""
-    c.run("poetry self update", echo=True, pty=True)
+    # c.run("poetry self update", echo=True, pty=True)
     c.run("poetry up --latest --only=dev,notebooks", echo=True, pty=True)
     c.run("poetry show --outdated --why", echo=True, pty=True)