+++ test/test_status.py
@@ -1,6 +1,6 @@
 from aio_overpass import Client
 from aio_overpass.client import Status
-from aio_overpass.error import ServerError
+from aio_overpass.error import ResponseError
 from test.util import VerifyingQueryRunner
 
 import pytest
@@ -205,7 +205,7 @@ open64: 2 No such file or directory /osm3s_osm_base Dispatcher_Client::1. Probab
 
     c = Client(runner=VerifyingQueryRunner())
 
-    with aioresponses() as m, pytest.raises(ServerError):
+    with aioresponses() as m, pytest.raises(ResponseError):
         m.get(
             url=URL_STATUS,
             body=body,